pyproject.toml
[tool.poetry]
name = "django-rest-framework_shoestore"
version = "0.1.0"
description = ""
authors = ["peyton glover <peyton.vx@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.11.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
